Author

+++ app/Http/Controllers/BackEnd/UserController.php
@@ -12,6 +12,8 @@ use Auth;
 use Illuminate\Support\Facades\File;
 use Illuminate\Support\Facades\Input;
 use Illuminate\Support\Facades\Session;
+use Gate;
+
 class UserController extends Controller
 {
     function __construct()
@@ -27,6 +29,10 @@ class UserController extends Controller
 
     public function index()
     {
+        if (Gate::denies('index-users')) {
+            abort(403);
+        }
+
         $users = User::orderBy('id','DESC')->paginate(15);
         return view('backend.users.index')->with('users',$users);
     }
@@ -38,6 +44,9 @@ class UserController extends Controller
      */
     public function create()
     {
+        if (Gate::denies('create-users')) {
+            abort(403);
+        }
         return view('backend.users.create');
     }
 
@@ -49,6 +58,9 @@ class UserController extends Controller
      */
     public function store(UserRequest $request)
     {
+        if (Gate::denies('create-users')) {
+            abort(403);
+        }
         $user = new User();
         $user->name = $request->name;
         $user->email = $request->email;
@@ -73,6 +85,10 @@ class UserController extends Controller
      */
     public function show($id)
     {
+        if (Gate::denies('show-users')) {
+            abort(403);
+        }
+
         $user = User::findOrFail($id);
         return view('backend.users.show')->with('user',$user);
     }
@@ -86,6 +102,11 @@ class UserController extends Controller
     public function edit($id)
     {
         $user = User::findOrFail($id);
+
+        if (Gate::denies('update-users', $user)) {
+            abort(403);
+        }
+
         return view('backend.users.edit')->with('user',$user);
     }
 
@@ -142,6 +163,9 @@ class UserController extends Controller
      */
     public function destroy($id)
     {
+        if (Gate::denies('delete-users')) {
+            abort(403);
+        }
         $user = User::findOrFail($id);
 
         Room::where('user_id', $id)->delete();
